+++ src/scripts/script.js
@@ -1,19 +1,3 @@
-function slowRange(tosave, time = 400) {
-	clearTimeout(rangeActive)
-	rangeActive = setTimeout(function () {
-		chrome.storage.sync.set(tosave)
-	}, time)
-}
-
-function slow(that, time = 400) {
-	that.setAttribute('disabled', '')
-	stillActive = setTimeout(() => {
-		that.removeAttribute('disabled')
-		clearTimeout(stillActive)
-		stillActive = false
-	}, time)
-}
-
 function traduction(settingsDom, lang = 'en') {
 	//
 	function traduis() {
@@ -43,7 +27,7 @@ function favicon(init, event) {
 
 	if (event) {
 		const val = event.value
-		const isEmoji = val.match(/\p{Emoji}/gu)
+		const isEmoji = val.match(/\p{Emoji}/gu) && !val.match(/[0-9a-z]/g)
 
 		if (isEmoji) createFavicon(val)
 		else event.value = ''
@@ -474,18 +458,20 @@ function quickLinks(event, that, initStorage) {
 		let touchStartTime = 0
 		let touchTimeout = setTimeout(() => {}, 0)
 
-		elem.ontouchstart = function (e) {
-			e.preventDefault()
+		const startHandler = () => {
 			touchStartTime = performance.now()
-			touchTimeout = setTimeout(() => editlink(this), 300)
+			touchTimeout = setTimeout(() => editlink(elem), 300)
 		}
 
-		elem.ontouchend = function (e) {
+		const endHandler = (e) => {
 			if (performance.now() - touchStartTime < 300) {
 				clearTimeout(touchTimeout)
-				openlink(this, e)
+				openlink(elem, e)
 			}
 		}
+
+		elem.addEventListener('touchstart', startHandler, { passive: true })
+		elem.addEventListener('touchend', endHandler, { passive: true })
 	}
 
 	function showDelIcon(input) {
@@ -495,42 +481,41 @@ function quickLinks(event, that, initStorage) {
 	}
 
 	function editEvents() {
+		const editLinkContainer = id('editlink_container')
+
 		function closeEditLink() {
 			removeLinkSelection()
-			id('edit_linkContainer').classList.add('hiding')
-			editDisplayTimeout = setTimeout(() => id('edit_linkContainer').setAttribute('class', ''), BonjourrAnimTime)
+			editLinkContainer.classList.add('hiding')
+			editDisplayTimeout = setTimeout(() => editLinkContainer.setAttribute('class', ''), BonjourrAnimTime)
 		}
 
 		function emptyAndHideIcon(e) {
 			e.target.previousElementSibling.value = ''
 			e.target.classList.remove('shown')
 		}
-
 		id('e_delete').onclick = function () {
 			removeLinkSelection()
-			removeblock(parseInt(id('edit_link').getAttribute('index')))
-			clas(id('edit_linkContainer'), false, 'shown')
+			removeblock(parseInt(id('editlink').getAttribute('index')))
+			clas(editLinkContainer, false, 'shown')
 			if (id('settings')) linksInputDisable(false)
 		}
 
 		id('e_submit').onclick = function () {
 			removeLinkSelection()
-			const noError = editlink(null, parseInt(id('edit_link').getAttribute('index')))
+			const noError = editlink(null, parseInt(id('editlink').getAttribute('index')))
 			if (noError) closeEditLink()
 		}
-
 		// close on button
 		id('e_close').onclick = () => closeEditLink()
 
 		// close on outside click
-		id('edit_linkContainer').onmousedown = (e) => {
-			if (e.target.id === 'edit_linkContainer') closeEditLink()
-		}
+		const outsideClick = (e) => (e.target.id === 'editlink_container' ? closeEditLink() : '')
+		if (mobilecheck) editLinkContainer.addEventListener('touchstart', outsideClick, { passive: true })
+		else editLinkContainer.onmousedown = outsideClick
 
 		id('re_title').onclick = (e) => emptyAndHideIcon(e)
 		id('re_url').onclick = (e) => emptyAndHideIcon(e)
 		id('re_iconurl').onclick = (e) => emptyAndHideIcon(e)
-
 		id('e_title').onkeyup = (e) => showDelIcon(e.target)
 		id('e_url').onkeyup = (e) => showDelIcon(e.target)
 		id('e_iconurl').onkeyup = (e) => showDelIcon(e.target)
@@ -612,14 +597,14 @@ function quickLinks(event, that, initStorage) {
 		else {
 			const index = findindex(that)
 			const liconwrap = that.querySelector('.l_icon_wrap')
-			const container = id('edit_linkContainer')
+			const container = id('editlink_container')
 			const openSettings = has(id('settings'), 'shown')
 
 			clas(liconwrap, true, 'selected')
 			clas(container, true, 'shown')
 			clas(container, openSettings, 'pushed')
 
-			id('edit_link').setAttribute('index', index)
+			id('editlink').setAttribute('index', index)
 
 			chrome.storage.sync.get('links', (data) => {
 				const { title, url, icon } = data.links[index]
@@ -745,12 +730,13 @@ function quickLinks(event, that, initStorage) {
 
 	function linksInputDisable(isMax, settingsDom) {
 		const getDoms = (name) => (settingsDom ? settingsDom.querySelector('#' + name) : id(name))
-		const doms = ['i_title', 'i_url', 'submitlink']
+		const doms = ['i_title', 'i_url', 'submitlink', 'b_importbookmarks']
 
 		if (isMax) doms.forEach((elem) => getDoms(elem).setAttribute('disabled', ''))
 		else doms.forEach((elem) => getDoms(elem).removeAttribute('disabled'))
 
 		clas(getDoms(doms[2]), isMax, 'max')
+		clas(getDoms(doms[3]), isMax, 'max')
 	}
 
 	switch (event) {
@@ -775,16 +761,21 @@ function quickLinks(event, that, initStorage) {
 
 		// No need to activate edit events asap
 		setTimeout(function timeToSetEditEvents() {
-			id('edit_linkContainer').oncontextmenu = (e) => e.preventDefault()
+			id('editlink_container').oncontextmenu = (e) => e.preventDefault()
 			editEvents()
 		}, 150)
 	}
 }
 
 async function linksImport() {
-	const changeCounter = (number) => (id('selectedCounter').textContent = `${number} / 30`)
+	const closeBookmarks = (container) => {
+		container.classList.add('hiding')
+		setTimeout(() => container.setAttribute('class', ''), BonjourrAnimTime)
+	}
 
 	function main(data, bookmarks) {
+		const changeCounter = (number) => (id('selectedCounter').textContent = `${number} / 30`)
+
 		const form = document.createElement('form')
 		const allCategories = [...bookmarks[0].children]
 		let counter = data.links.length || 0
@@ -793,7 +784,6 @@ async function linksImport() {
 
 		changeCounter(counter)
 		allCategories.forEach((cat) => bookmarksList.push(...cat.children))
-		console.log(...bookmarksList)
 
 		bookmarksList.forEach((mark, index) => {
 			const elem = document.createElement('div')
@@ -808,34 +798,48 @@ async function linksImport() {
 			elem.appendChild(url)
 			elem.onclick = () => {
 				const isSelected = elem.classList.toggle('selected')
+				const isFull = isSelected && counter === 30
 
-				if (isSelected && counter === 30) elem.classList.toggle('selected')
+				// Color counter
+				clas(id('selectedCounter'), isSelected && counter > 28, 'full')
+
+				// unselect selection if full
+				if (isFull) elem.classList.toggle('selected')
 				else {
 					isSelected ? selectedList.push(elem.getAttribute('index')) : selectedList.pop()
 					isSelected ? counter++ : (counter -= 1)
 					changeCounter(counter)
 				}
 
-				clas(id('applybookmarks'), counter > (data.links.length || 0), 'shown')
+				// Change submit button text & class on selections
+				const amountSelected = counter - data.links.length
+				id('applybookmarks').textContent = tradThis(
+					amountSelected === 0
+						? 'Select bookmarks to import'
+						: amountSelected === 1
+						? 'Import this bookmark'
+						: 'Import these bookmarks'
+				)
+				clas(id('applybookmarks'), amountSelected === 0, 'none')
 			}
 
+			// only append links if url are not empty
+			// (temp fix to prevent adding bookmarks folder title ?)
 			if (typeof mark.url === 'string')
 				if (data.links.filter((x) => x.url === stringMaxSize(mark.url, 128)).length === 0) form.appendChild(elem)
 		})
 
+		// Replace form to filter already added bookmarks
 		const oldForm = document.querySelector('#bookmarks form')
 		if (oldForm) oldForm.remove()
-
-		// id('bookmarks').appendChild(form)
-		// id('e_close').after(form)
 		id('bookmarks').insertBefore(form, document.querySelector('#bookmarks .bookmarkOptions'))
 
+		// Submit event
 		id('applybookmarks').onclick = function () {
 			const bookmarkToApply = selectedList.map((i) => ({ title: bookmarksList[i].title, url: bookmarksList[i].url }))
-			console.log(...bookmarkToApply)
 
 			if (bookmarkToApply.length > 0) {
-				id('bookmarks').style.display = 'none'
+				closeBookmarks(id('bookmarks_container'))
 				quickLinks('button', bookmarkToApply, null)
 			}
 		}
@@ -846,11 +850,18 @@ async function linksImport() {
 		if (granted) {
 			chrome.storage.sync.get('links', (data) => {
 				;(window.location.protocol === 'moz-extension:' ? browser : chrome).bookmarks.getTree().then((response) => {
-					id('bookmarks_container').style.display = 'flex'
+					clas(id('bookmarks_container'), true, 'shown')
 					main(data, response)
 				})
 			})
 		}
+	})
+
+	// Close events
+	document.querySelector('#bookmarks #e_close').onclick = () => closeBookmarks(id('bookmarks_container'))
+
+	id('bookmarks_container').addEventListener('click', function (e) {
+		if (e.target.id === 'bookmarks_container') closeBookmarks(this)
 	})
 }
 
@@ -884,32 +895,45 @@ function weather(event, that, init) {
 		'N2M1NDFjYWVmNWZjNzQ2N2ZjNzI2N2UyZjc1NjQ5YTk=',
 	]
 
-	function initWeather(param) {
-		navigator.geolocation.getCurrentPosition(
-			(pos) => {
-				//update le parametre de location
-				param.location.push(pos.coords.latitude, pos.coords.longitude)
-				request(param, false)
-				request(param, true)
+	async function initWeather(param) {
+		const applyResult = (geol) => {
+			request(param, true)
+			request(param, false)
 
-				// Check geolocation after settings is loaded
-				if (id('settings')) {
+			if (id('settings')) {
+				id('i_ccode').value = param.ccode
+				id('i_city').setAttribute('placeholder', param.city)
+
+				if (geol) {
 					clas(id('sett_city'), true, 'hidden')
 					id('i_geol').checked = true
 				}
-			},
-			(refused) => {
-				request(param, true)
-				request(param, false)
 			}
+		}
+
+		try {
+			const ipapi = await fetch('https://ipapi.co/json')
+			if (ipapi.ok) {
+				const json = await ipapi.json()
+				if (!json.error) param = { ...param, city: json.city, ccode: json.country }
+			}
+		} catch (error) {
+			console.warn(error)
+		}
+
+		navigator.geolocation.getCurrentPosition(
+			(pos) => {
+				param.location.push(pos.coords.latitude, pos.coords.longitude)
+				applyResult(true)
+			},
+			() => applyResult(false)
 		)
 	}
 
-	function request(storage, forecast) {
+	async function request(storage, forecast) {
 		function saveCurrent(response) {
 			//
 			const isImperial = storage.unit === 'imperial'
-
 			weatherToSave = {
 				...weatherToSave,
 				lastCall: Math.floor(new Date().getTime() / 1000),
@@ -922,14 +946,11 @@ function weather(event, that, init) {
 					icon_id: response.weather[0].id,
 				},
 			}
-
 			chrome.storage.sync.set({ weather: weatherToSave })
 			displaysCurrent(weatherToSave)
 		}
-
 		function saveForecast(response) {
 			//
-
 			const thisdate = new Date()
 			const todayHour = thisdate.getHours()
 			let forecastDay = thisdate.getDate()
@@ -948,32 +969,35 @@ function weather(event, that, init) {
 			})
 
 			weatherToSave.fcHigh = Math.round(storage.unit === 'imperial' ? toFarenheit(maxTempFromList) : maxTempFromList)
-
 			chrome.storage.sync.set({ weather: weatherToSave })
 			displaysForecast(weatherToSave)
 		}
-
 		let url = 'https://api.openweathermap.org/data/2.5/'
 		const lang = document.documentElement.getAttribute('lang')
 		const [lat, lon] = storage.location || [0, 0]
-
 		url += `${forecast ? 'forecast' : 'weather'}?appid=${atob(WEATHER_API_KEY[forecast ? 0 : 1])}`
 		url += storage.location.length === 2 ? `&lat=${lat}&lon=${lon}` : `&q=${encodeURI(storage.city)},${storage.ccode}`
 		url += `&units=metric&lang=${lang}`
-
 		// Inits global object
 		if (Object.keys(weatherToSave).length === 0) {
 			weatherToSave = storage
 		}
 
 		// fetches, parses and apply callback
-		fetch(url).then((data) => {
-			if (data.ok) data.json().then((json) => (forecast ? saveForecast(json) : saveCurrent(json)))
-			else if (i_city) i_city.setAttribute('placeholder', tradThis('City not found'))
-		})
+		try {
+			const weatherAPI = await fetch(url)
+
+			if (weatherAPI.ok) {
+				const json = await weatherAPI.json()
+				forecast ? saveForecast(json) : saveCurrent(json)
+			}
+			return weatherAPI.ok
+		} catch (error) {
+			return false
+		}
 	}
 
-	function cacheControl(storage) {
+	function weatherCacheControl(storage) {
 		const now = Math.floor(date.getTime() / 1000)
 		let isCurrentChanging = false
 
@@ -1064,11 +1088,17 @@ function weather(event, that, init) {
 		const timeOfDay = now < rise || now > set ? 'night' : 'day'
 		const iconSrc = `src/assets/weather/${timeOfDay}/${filename}.png`
 
-		const icon = document.createElement('img')
-		icon.src = iconSrc
-		icon.setAttribute('draggable', 'false')
+		if (widgetIcon) {
+			if (widgetIcon.getAttribute('src') !== iconSrc) widgetIcon.setAttribute('src', iconSrc)
+		} else {
+			const icon = document.createElement('img')
+			icon.src = iconSrc
+			icon.setAttribute('draggable', 'false')
+			widget.prepend(icon)
 
-		!widgetIcon ? widget.prepend(icon) : widgetIcon.setAttribute('src', iconSrc)
+			// from 1.2s request anim to .4s hide elem anim
+			setTimeout(() => (widget.style.transition = 'opacity .4s'), BonjourrAnimTime)
+		}
 
 		// Description
 		const desc = weather.lastState.description
@@ -1079,87 +1109,113 @@ function weather(event, that, init) {
 
 		clas(current, false, 'wait')
 		clas(widget, false, 'wait')
-
-		// from 1.2s request anim to .4s hide elem anim
-		setTimeout(() => (widget.style.transition = 'opacity .4s'), BonjourrAnimTime)
 	}
 
 	function displaysForecast(weather) {
-		const when = tradThis(date.getHours() > 21 ? 'tomorrow' : 'today')
+		forecast.textContent = `${tradThis('with a high of')} ${weather.fcHigh}° ${tradThis(
+			date.getHours() > 21 ? 'tomorrow' : 'today'
+		)}.`
 
-		forecast.textContent = `${tradThis('with a high of')} ${weather.fcHigh}° ${when}.`
 		clas(forecast, false, 'wait')
 	}
 
-	function updatesWeather() {
+	function forecastVisibilityControl(value) {
+		let isTimeForForecast = false
+
+		if (value === 'auto') isTimeForForecast = date.getHours() < 12 || date.getHours() > 21
+		else isTimeForForecast = value === 'always'
+
+		clas(forecast, isTimeForForecast, 'shown')
+	}
+
+	async function updatesWeather() {
 		//
 
-		function fetches(weather) {
-			request(weather, false)
-			request(weather, true)
+		async function fetches(weather) {
+			const main = await request(weather, false)
+			const forecast = await request(weather, true)
+
+			return main && forecast
 		}
 
-		slow(that)
-
-		chrome.storage.sync.get('weather', (data) => {
+		chrome.storage.sync.get('weather', async (data) => {
 			switch (event) {
 				case 'units': {
 					data.weather.unit = that.checked ? 'imperial' : 'metric'
-
 					if (data.weather.lastState) {
 						const { feels_like } = data.weather.lastState
 						data.weather.lastState.feels_like = that.checked ? toFarenheit(feels_like) : toCelsius(feels_like)
 						data.weather.fcHigh = that.checked ? toFarenheit(data.weather.fcHigh) : toCelsius(data.weather.fcHigh)
 					}
-
 					displaysCurrent(data.weather)
 					displaysForecast(data.weather)
 					chrome.storage.sync.set({ weather: data.weather })
+					slow(that)
 					break
 				}
 
 				case 'city': {
-					if (i_city.value.length < 2) {
-						return false
+					slow(that)
+
+					if (i_city.value.length < 3) return false
+					else if (navigator.onLine) {
+						data.weather.ccode = i_ccode.value
+						data.weather.city = stringMaxSize(i_city.value, 64)
+
+						const inputAnim = i_city.animate([{ opacity: 1 }, { opacity: 0.6 }], {
+							direction: 'alternate',
+							easing: 'linear',
+							duration: 800,
+							iterations: Infinity,
+						})
+
+						const cityFound = await fetches(data.weather)
+
+						if (cityFound) i_city.blur()
+						i_city.setAttribute('placeholder', cityFound ? data.weather.city : tradThis('City not found'))
+						i_city.value = ''
+						inputAnim.cancel()
 					}
 
-					data.weather.ccode = i_ccode.value
-					data.weather.city = stringMaxSize(i_city.value, 64)
-
-					fetches(data.weather)
-
-					i_city.setAttribute('placeholder', data.weather.city)
-					i_city.value = ''
-					i_city.blur()
 					break
 				}
 
 				case 'geol': {
 					data.weather.location = []
-					clas(sett_city, that.checked, 'hidden')
+					that.setAttribute('disabled', '')
 
 					if (that.checked) {
-						that.setAttribute('disabled', '')
-
 						navigator.geolocation.getCurrentPosition(
 							(pos) => {
 								//update le parametre de location
+								clas(sett_city, that.checked, 'hidden')
 								data.weather.location.push(pos.coords.latitude, pos.coords.longitude)
 								fetches(data.weather)
 							},
 							(refused) => {
 								//désactive geolocation if refused
-								that.checked = false
+								setTimeout(() => (that.checked = false), 400)
 								if (!data.weather.city) initWeather()
+								console.log(refused)
 							}
 						)
 					} else {
 						i_city.setAttribute('placeholder', data.weather.city)
 						i_ccode.value = data.weather.ccode
+						clas(sett_city, that.checked, 'hidden')
 
 						data.weather.location = []
 						fetches(data.weather)
 					}
+
+					slow(that)
+					break
+				}
+
+				case 'forecast': {
+					data.weather.forecast = that.value
+					chrome.storage.sync.set({ weather: data.weather })
+					forecastVisibilityControl(that.value)
 					break
 				}
 			}
@@ -1168,16 +1224,12 @@ function weather(event, that, init) {
 
 	// Event & Init
 	if (event) updatesWeather()
-	else cacheControl(init)
-
-	// Detect forecast display before it fetches
-	const isTimeForForecast = date.getHours() < 12 || date.getHours() > 21
-	clas(forecast, isTimeForForecast, 'shown')
-
-	// Checks every 5 minutes if weather needs update
-	setTimeout(() => {
-		navigator.onLine ? chrome.storage.sync.get(['weather'], (data) => cacheControl(data.weather)) : ''
-	}, 5 * 60 * 1000)
+	else {
+		if (typeof init.hide === 'object' && init.hide[1][1] + init.hide[1][2] < 2) {
+			forecastVisibilityControl(init.weather.forecast || 'mornings')
+			weatherCacheControl(init.weather)
+		}
+	}
 }
 
 function initBackground(data) {
@@ -1209,7 +1261,7 @@ function initBackground(data) {
 	filter('init', [parseFloat(blur), parseFloat(bright)])
 }
 
-function imgBackground(val, loadTime) {
+function imgBackground(val, loadTime, init) {
 	let img = new Image()
 
 	img.onload = () => {
@@ -1218,11 +1270,15 @@ function imgBackground(val, loadTime) {
 			const changeDuration = (time) => (domoverlay.style.transition = `transform .4s, opacity ${time}ms`)
 
 			changeDuration(animDuration)
-			setTimeout(() => changeDuration(400), animDuration)
+			setTimeout(() => changeDuration(BonjourrAnimTime), animDuration)
 		}
 
-		domoverlay.style.opacity = `1`
-		id('background').style.backgroundImage = `url(${val})`
+		const applyBackground = () => {
+			domoverlay.style.opacity = `1`
+			id('background').style.backgroundImage = `url(${val})`
+		}
+
+		init ? applyBackground() : setTimeout(applyBackground, BonjourrAnimTime)
 	}
 
 	img.src = val
@@ -1274,9 +1330,10 @@ function localBackgrounds(init, event) {
 		const background = backgrounds[index]
 
 		if (background) {
+			const perfStart = performance.now()
 			const cleanData = background.slice(background.indexOf(',') + 1, background.length)
 			b64toBlobUrl(cleanData, (bloburl) => {
-				imgBackground(bloburl)
+				imgBackground(bloburl, perfStart, !!init)
 				changeImgIndex(index)
 			})
 		}
@@ -1330,9 +1387,8 @@ function localBackgrounds(init, event) {
 		//
 		// Hides previous bg and credits
 		if (state !== 'thumbnail') {
-			domoverlay.style.opacity = `0`
 			clas(domcredit, false, 'shown')
-			setTimeout(() => (domcredit.style.display = 'none'), BonjourrAnimTime)
+			domoverlay.style.opacity = `0`
 		}
 
 		const compressStart = performance.now()
@@ -1453,12 +1509,9 @@ function localBackgrounds(init, event) {
 					// Last image is removed
 					if (data.custom.length === 0) {
 						domoverlay.style.opacity = `0`
-						domcredit.style.display = 'block'
 
-						setTimeout(() => {
-							unsplash(null, { removedCustom: true })
-							clas(domcredit, true, 'shown')
-						}, 400)
+						unsplash(null, { removedCustom: true })
+						clas(domcredit, true, 'shown')
 					}
 
 					// Only draw new image if displayed is removed
@@ -1538,7 +1591,6 @@ function unsplash(init, event) {
 		//
 		const country = image.country || 'Photo'
 		const city = image.city ? image.city + ', ' : ''
-
 		const credits = [
 			{
 				text: city + country,
@@ -1554,22 +1606,22 @@ function unsplash(init, event) {
 			},
 		]
 
-		id('credit').textContent = ''
+		domcredit.textContent = ''
 
 		credits.forEach(function cityNameRef(elem, i) {
 			const dom = document.createElement('a')
 			dom.textContent = elem.text
 			dom.href = elem.url
 
-			if (i === 1) id('credit').appendChild(document.createElement('br'))
-			id('credit').appendChild(dom)
+			if (i === 1) domcredit.appendChild(document.createElement('br'))
+			domcredit.appendChild(dom)
 		})
 
-		clas(id('credit'), true, 'shown')
+		clas(domcredit, true, 'shown')
 	}
 
 	function loadBackground(props, loadTime) {
-		imgBackground(props.url, loadTime)
+		imgBackground(props.url, loadTime, !!init)
 		imgCredits(props)
 
 		// sets meta theme-color to main background's color
@@ -1665,7 +1717,10 @@ function unsplash(init, event) {
 			if (list.length === 1)
 				requestNewList(collection, (newlist) => {
 					caches[collection] = list.concat(newlist)
-					noDisplayImgLoad(newlist[0].url, () => chrome.storage.local.set({ dynamicCache: caches }))
+					noDisplayImgLoad(newlist[0].url, () => {
+						chrome.storage.local.set({ dynamicCache: caches })
+						chrome.storage.local.remove('waitingForPreload')
+					})
 				})
 			//
 			// Or preload next
@@ -1673,6 +1728,7 @@ function unsplash(init, event) {
 				noDisplayImgLoad(list[1].url, () => {
 					chrome.storage.sync.set({ dynamic: dynamic })
 					chrome.storage.local.set({ dynamicCache: caches })
+					chrome.storage.local.remove('waitingForPreload')
 				})
 		}
 
@@ -1756,10 +1812,42 @@ function unsplash(init, event) {
 
 		case 'event': {
 			chrome.storage.sync.get('dynamic', (data) => {
-				chrome.storage.local.get('dynamicCache', (local) => {
+				chrome.storage.local.get(['dynamicCache', 'waitingForPreload'], (local) => {
 					//
 
 					switch (Object.keys(event)[0]) {
+						case 'refresh': {
+							const buttonSpan = Object.values(event)[0]
+							const animationOptions = { duration: 600, easing: 'ease-out' }
+
+							// Only refreshes background if preload is over
+							// If not, animate button to show it is trying
+							if (local.waitingForPreload === undefined) {
+								id('background_overlay').style.opacity = 0
+								data.dynamic.time = 0
+								chrome.storage.sync.set({ dynamic: data.dynamic })
+								chrome.storage.local.set({ waitingForPreload: true })
+
+								buttonSpan.animate([{ transform: 'rotate(360deg)' }], animationOptions)
+
+								setTimeout(
+									() =>
+										cacheControl(data.dynamic, local.dynamicCache, collectionControl(data.dynamic), false),
+									BonjourrAnimTime
+								)
+							} else
+								buttonSpan.animate(
+									[
+										{ transform: 'rotate(0deg)' },
+										{ transform: 'rotate(90deg)' },
+										{ transform: 'rotate(0deg)' },
+									],
+									animationOptions
+								)
+
+							break
+						}
+
 						case 'every': {
 							data.dynamic.every = event.every
 							data.dynamic.time = freqControl('set')
@@ -1956,11 +2044,10 @@ function searchbar(event, that, init) {
 			const request = domsearchbar.getAttribute('request')
 
 			// engineLocales est dans lang.js
-
 			if (engine === 'custom') searchURL = request
 			else searchURL = engineLocales[engine].base.replace('%l', engineLocales[engine][lang])
 
-			searchURL = searchURL.replace('%s', this.value)
+			searchURL = searchURL.replace('%s', encodeURIComponent(this.value))
 
 			isNewtab ? window.open(searchURL, '_blank') : (window.location = searchURL)
 		}
@@ -2036,10 +2123,7 @@ function showPopup(data) {
 function customSize(init, event) {
 	//
 	// Apply for interface, credit & settings button
-	const apply = (size) => {
-		dominterface.style.fontSize = size + 'px'
-		id('credit').style.fontSize = size + 'px'
-	}
+	const apply = (size) => (dominterface.style.fontSize = size + 'px')
 
 	const save = () => {
 		chrome.storage.sync.get('font', (data) => {
@@ -2304,12 +2388,10 @@ function hideElem(init, buttons, that) {
 	]
 
 	// Returns { row, col } to naviguate [[0, 0], [0, 0, 0]] etc.
-	function getEventListPosition(that) {
-		return {
-			row: parseInt(that.getAttribute('he_row')),
-			col: parseInt(that.getAttribute('he_col')),
-		}
-	}
+	const getEventListPosition = (that) => ({
+		row: parseInt(that.getAttribute('he_row')),
+		col: parseInt(that.getAttribute('he_col')),
+	})
 
 	function toggleElement(dom, hide) {
 		if (hide) id(dom).classList.add('he_hidden')
@@ -2336,57 +2418,14 @@ function hideElem(init, buttons, that) {
 		})
 	}
 
-	function updateToNewData(list) {
-		if (list[0]) {
-			if (typeof list[0][0] === 'string') {
-				//
-				// Flattens and removes parent IDs
-				const childOnly = IDsList.flat().filter((row) => typeof row === 'object')
-				let newHidden = [[0, 0], [0, 0, 0], [0], [0]]
-
-				//
-				// Go through IDs list for every old hide elems
-				list.forEach((id) => {
-					childOnly.forEach((row, row_i) =>
-						row.forEach((col, col_i) => {
-							if (col === id) {
-								newHidden[row_i][col_i] = 1
-							}
-						})
-					)
-				})
-
-				chrome.storage.sync.set({ hide: newHidden })
-				return newHidden
-			}
-
-			// Is already updated
-			else return list
-		}
-
-		// 1.10.0 online hotfix
-		if (list.length === 0) {
-			let newHidden = [[0, 0], [0, 0, 0], [0], [0]]
-			chrome.storage.sync.set({ hide: newHidden })
-			return newHidden
-		}
-
-		// Had nothing to hide
-		else return list
-	}
-
 	// startup initialization
 	if (!that && !buttons) {
-		initializeHiddenElements(updateToNewData(init))
+		initializeHiddenElements(init)
 	}
 
 	// Settings buttons initialization
 	else if (buttons) {
 		chrome.storage.sync.get('hide', (data) => {
-			//
-			// 1.9.3 ==> 1.10.0
-			data.hide = updateToNewData(data.hide)
-
 			buttons.forEach((button) => {
 				const pos = getEventListPosition(button)
 				if (data.hide[pos.row][pos.col] === 1) button.classList.toggle('clicked')
@@ -2396,11 +2435,7 @@ function hideElem(init, buttons, that) {
 
 	// Event
 	else {
-		chrome.storage.sync.get('hide', (data) => {
-			//
-			// 1.9.3 ==> 1.10.0
-			data.hide = updateToNewData(data.hide)
-
+		chrome.storage.sync.get(['weather', 'hide'], (data) => {
 			const pos = getEventListPosition(that)
 			const state = that.classList.contains('clicked')
 			const child = IDsList[pos.row][1][pos.col]
@@ -2409,6 +2444,9 @@ function hideElem(init, buttons, that) {
 			// Update hidden list
 			data.hide[pos.row][pos.col] = state ? 1 : 0
 			chrome.storage.sync.set({ hide: data.hide })
+
+			// Re-activates weather
+			if (!state && pos.row === 1 && pos.col > 0) weather(null, null, data)
 
 			// Toggle children and parent if needed
 			toggleElement(child, state)
@@ -2440,6 +2478,7 @@ function canDisplayInterface(cat, init) {
 
 		// Prevent error message from appearing
 		clearTimeout(errorMessageInterval)
+		if (id('bonjourrError')) id('bonjourrError').remove()
 	}
 
 	// More conditions if user is using advanced features
@@ -2541,6 +2580,7 @@ function filterImports(data) {
 
 				if (weather.lastCall) weather.lastCall = 0
 				if (weather.forecastLastCall) delete weather.forecastLastCall
+				if (weather.forecast === undefined) weather.forecast = 'auto'
 			}
 
 			return weather
@@ -2589,7 +2629,7 @@ function startup(data) {
 	traduction(null, data.lang)
 
 	sunTime(data.weather)
-	weather(null, null, data.weather)
+	weather(null, null, data)
 
 	customFont(data.font)
 	customSize(data.font)
@@ -2610,6 +2650,27 @@ function startup(data) {
 }
 
 window.onload = function () {
+	if (mobilecheck) {
+		// For Mobile that caches pages for days
+		document.addEventListener('visibilitychange', () => {
+			chrome.storage.sync.get(['dynamic', 'waitingForPreload', 'weather', 'background_type', 'hide'], (data) => {
+				const { dynamic, background_type } = data
+				const dynamicNeedsImage = background_type === 'dynamic' && freqControl('get', dynamic.every, dynamic.time)
+
+				if (dynamicNeedsImage) {
+					domoverlay.style.opacity = 0
+					unsplash(data, false)
+				}
+
+				weather(null, null, data)
+			})
+		})
+	}
+
+	// Checks every 5 minutes if weather needs update
+	setInterval(() => {
+		navigator.onLine ? chrome.storage.sync.get(['weather', 'hide'], (data) => weather(null, null, data)) : ''
+	}, 5 * 60 * 1000)
 	//
 	// Only on Online
 	switch (window.location.protocol) {
@@ -2642,7 +2703,7 @@ window.onload = function () {
 		chrome.storage.sync.get(null, (data) => {
 			errorMessageInterval = setTimeout(() => {
 				errorMessage(JSON.stringify(data), null)
-			}, 1000)
+			}, 2000)
 
 			//
 			let newVersion = !data.about